implementation of stepover / stepinto
StepOver will step over a given line. If the line contains a function the function will be executed and the result returned without debugging each line.
StepInto If the line does not contain a function it behaves the same as “step over” but if it does the debugger will enter the called function and continue line-by-line debugging there.